(PUP-6480) Improve error message for string[string]
Before this, if a user forgot to capitalize a type when writing a
reference, for example `user[fred]` - this would result in "cannot
convert 'fred' to numeric".

This was not the intended behavior as the specially design issue code
for this was never issued as the string to numeric conversion method
used also asserted the result.

Now, the output is "A substring operation does not accept a String as a
character index. Expected an Integer". The error also points to the char
index value in question.

As this mentions "substring operation", it should be much clearer what
the parser interprets the source text to mean.