(doc) Fix empty example and parameter types in regsubst function.
The `regsubst` function had an empty @example tag which was causing Yard to
fail to render documentation for the function.  Removing the empty example and
replacing with the correct examples following the parameter documentation.

The parameter types in the @param tags did not match those specified in the
dispatch entries, causing `puppet strings` to emit warnings for the mismatch.
This commit updates the types to match and uses the Yardoc supported separator
of ',' instead of '|'.