(PUP-6338) Move 3.x contain() to 4.x API and make it return references
This refactors the 3.x contain() function to make it use the 4.x API.
As requested in PUP-6338 this also makes the function return an array
containing Class[] references to the class names (in different forms)
given to the function.

The 3.x version of the function is retained as a stub for documentation
generation purposes.