(PUP-6338) Refactor function require to 4.x and returns refs
Before this, the require function was a 3.x statement function that
could not return a value.
Now it is migrated to the 4.x function API and returns an
Array[Type[Class]] with the classes required by the function.

As a consequence of this change, the spec tests were updated (to a more
modern an easy to read form), and reduntant integration tests were
removed.

The shared tests for all functions that deal with class references (the
family: include, require, contain) needed an update to also support the
require function as it checks for the existence of the class that it is
using in the tests.