(PUP-6103) Make the epp render command use facts from local node
Before this, a user would always have to use --facts factsfile to get
facts set when rendering an epp termplate.

Now, facts from the local node will always be used, or from a given node
if using --node <nodename>.
In addition, if --facts <factfile> is given, it may now be in either
.yaml, or .json as indicated by the suffix of the file, and the content
of the --facts file will override facts set in the node.