(PUP-6629) Respect the passed log level when log_exception is passed a custom message
The first time we tried to fix PUP-6629, we did not fix all the code
paths in log_exception - specifically, when a custom message was
passed (which insync_values? does), we still logged at error level.

This now handles that case, and also adds a test for insync_values? to
ensure it does not log at error level if insync? fails.