(PUP-5871) Add Timespan and Timestamp Puppet Types and implementations
This commit adds two new types. A Timespan that represents a moment
in time and a Timespan that represents a duration. Both types uses
nano-second precision and can participate in arithmetic operations.

The types have `new` functions that allow creation from a string a hash.

A strftime function is also included to provide a flexible way of
representing the instances in string form. The name was chosen since
it's well known in languages such as Ruby, C++, and Python. It's also
less likely to collide with other global functions declared in modules
than, say 'format'.

An approach of extending the String constructor and give it special
logic for Timespan and Timestamp was rejected since it would introduce
ambiguities with different formatting characters. `String(x, '%p')`
(expected to produce a programmatic representation) would print 'AM' or
'PM' in case `x` is a Timestamp.

The Timespan and Timestamp implementation share an abstract class
TimeData which in turn inherits from Numeric. This approach was selected
since it required a minimum of adjustments to make the classes work
with arithmetic and comparison operators in PL.