(PUP-6659) Remove Stacktraces in API responses
Previously, Internal Server Errors would return an API response that
included a stacktrace.

This is not considered a best practice and on non-standard setups
filesystem locations disclosed in a stacktrace could be considered
sensitive information.

This commit moves to logging the stacktrace at the error level while
returning only the exception message via the HTTP API.