(PUP-6592) acceptance: create table driven resource testing prototype
In this commit we create tools to test resources which contain their own
assertions, parameters/properties and expect_failures.
The tooling produces a single manifest from an array of hashes
containing the above information.  It uses the same array of hashes to
produce a string of ruby assertion-laden code to run against the result
of beaker's `#on` or a string. There are also spec tests.

[skip ci]