(maint) Fix problem with arithmetic involving TimeData on right side
This commit fixes problems that occurs when a Timestamp or Timespan
is on the right hand side of an arithmetic operation and the left
hand side is an Integer or Float. The result of an addition,
subtraction, or multiplication was of the wrong type and some
nonsensical operations that really should be prohibited, were allowed.

A side effect of this fix is an improvement of the error message
when using '%' with a Float on the right hand side. The expression

1 % 1.0

would, before this commit, result in the message:

Operator '%' is not applicable to an Integer

now, the message will instead be:

Operator '%' is not applicable to an Integer when right side is a Float