(PUP-6663) quote query format strings in rpm provider to prevent stale locks on the rpm database
When running large volumes of rpm queries as the root user to determine package state, locks are placed on the rpm database that can become stale and prevent package installation/removal through either the rpm or yum providers.

The fix to prevent stale locks from occurring in situations of large query volumes appears to be properly quoting the query format string (represented as NEVRA_FORMAT) in the rpm provider.