(PUP-6662) Make Sensitive data work in user defined resources
Before this, a Sensitive type value could only be used directly
in a native resource.

In order to fix this, there are some hoops to jump through.
* a puppet apply does not serialize the resources in the catalog which
means that the processing of sensitive data must take place in the
resources themselves.
* The early unwrapping and marking parameters as sensitive then occur
too early (finish) for classes and user defined types. For native types
there is no evaluation, and finish is only called post compile, but for
user defined resources / classes this occurs when the resource is
evaluated.