(PUP-4643) Add evaluation control functions break(), return() and next()
This adds the three evaluation control functions to puppet:
* break() breaks iteration as if the current iteration reached end of
input.
* return(v) returns from a function, class or define body
* next(v) returns from a code block, or from a function, class or define
body

The implementation is made complicated by the fact that it is a function
that originates break/return/next and the function generating this
semantic action is clearly not the function that the break/return/next
applies to. Secondly, a StopIteration exception must be used for break
since it needs to be able to break a Ruby enumeration. Thirdly, code
blocks and named functions share most of their implementation and they
need to treat return() and next() differently.

A final difficulty was handling break() from contexts where this is
illegal - catching the StopIteration and reporting the location from
which it was called.

This commit contains solutions to all of these.