(PUP-6670) Update tag validation regex to handle newlines
This patch updates the tag validation regex so that it is anchored by `\A` and
`\Z` instead of `^` and `$`. The issue with using `$` is that it will terminate
validation when a newline is encountered, which allows non-valid tags to pass
through, which causes catalogs to be rejected by PuppetDB. The updated pattern
matches the anchors PuppetDB uses to validate tags.