(maint) Make it possible to use a StandardError as semantic in issue
Before this, it was required to send a model object, or a
Puppet::Resource as the semantic object in an issue based error.

This meant that if a raised error was the carrier of the semantic
information (file, line, pos) code would have to jump through hoops to
make use of that information via some other semantic object.

Now, a StandardError subclass that has the methods #file(), #line() and
#pos() can be used as a semantic object.