(maint) Make time data compare equal to numbers with the same magnitude
Before this commit, a `Timespan` or `Timestamp` would yield `true` when
the operator `>=` or `<=` was used to compare them with an `Integer` or
`Float` that had the same magnitude (seconds or seconds since epoch) but
yield `false` when using the operator `==` on the same operands. This
commit changes this so that the latter case also yields `true`.