(PUP-6641) Fix Service integration test for Windows 10
Previously the integration test for the Service provider for Windows used the
lmhosts service as a test fixture.  However in Windows 10 the start type for
this service changed to Manual (Triggered) which now fails the integration test.

This commit adds `:manual` as a valid return value of `enabled?`, which now
matches the provider for service in Windows.