(PUP-6701) Make settings data type cleanup handle boolean values
A regression was introduced in PUP-6682 as the transformation of
settings values to safe values (no symbols) omitted handling boolean
values. They ended up being transformed to string, with the unfortunate
effect that a boolean `false` turned into the string `"true"` which is
"truthy".

This commit adds the correct handling of FalseClass and TrueType so that
they are not transformed to String values.