(FACT-1497) Extend variables_refreshed sleep on AIX
This commit extends the `sleep` statement in the
`variables_refreshed_each_compilation` test to 61 seconds when
the platform is AIX.

As per FACT-1497, the `system_uptime.seconds` fact updates
less frequently on AIX. Prior to this change this test failed
on AIX due to this issue with facter.

This is a work around and should be reverted when the underlying
facter issue is resolved.

[skip-ci]