(PUP-6703) Tweak failover_master test for robustness
On 9/14 we had a failed cell (spurious failure) in the matrix run for test
acceptance/tests/reports/failover_master.rb due to a timeout sending the report
to the master for the test modified in this commit, 'master_field should not
appear when no master could be contacted'.

Prior to this commit, this test used two nonexistent servers in a server list
and specified a timeout of 3 seconds. In the beaker run noted, this led to a
test execution time of approximately 25 seconds. This commit updates the
timeout to 5 seconds, but reduces the check to only one nonexistent master as a
way to diminish the impact on overall test time of increasing the timeout.  The
fact that two masters are specified isn't actually intrinsic to the test, so
this should effectively be a no-op in terms of test effectiveness while
decreasing the likelihood of encountering a timeout during report submission.