(PUP-5831) Improve docs and remove useless test
This removes a test that did not assert what it claimed. The test is to
be replaced later by adding the Binary type to the "types_setup.rb" to
make it part of the more extensive testing of each type against all
others. That cannot be done now as Pcore serialization support is needed
and that is going to be added later.

The documentation improvements point out the rules for %#s format.