(PUP-5834) Change Runtime3Converter to allow specializations
Before this commit, the Runtime3Converter relied on that all convert_XXX
methods where aliased with a corresponding convert2_XXX method to cater
for different mappings for array and hash content. This prohibited
specializations of the class since an alias isn't inherited.

This commit removes the aliases in favor of using two instances of the
class where an instance attribute named @inner controls the different
behavior for the two methods in question `convert_NilClass`and
`convert_Symbol`.

The reason for this change is that subsequent changes will require
different behavior when converting function arguments for a 3x function
and resource parameters.