(PUP-5834) Retain 3.x function argument conversion
Before this commit, the `Runtime3Converter` would always convert the
`Regexp`, `Version`, `VersionRange`, `Timespan`, and `Timestamp` classes
to strings. This behavior is however only desired when converting
arguments that are passed to 3.x functions. The conversion must not
happen when converting resource parameter values.

This commit creates the subclass `Runtime3FunctionArgumentConverter`
and moves the 3x function argument specific conversions to that class.
That class is then used in place of its superclass when 3x function
calls are made.