(PUP-5834) Enable access to reader/writer used by serialization
This commit adds methods that enables the internal reader/writer
used by the Puppet Serialization to be accessed and re-initialized.
The methods are marked `@api private` and intended to be used only
when one or more fragments of a larger JSON document will be
serialized/deserialized using the same tabulation context.