(PUP-5834) Add support for non-Data parameter values in the catalog
This commit adds a new `ext_parameters` entry to the JSON object
that represents a Resource. The `ext_parameters`, like `parameters`
is an object keyed by parameter names. The values are however not
plain JSON types but instead an array representing an extended type
created by the Puppet Serialization API.

Special hooks into the serializer enables several values to be
generated, one at a time, while using the same tabulation context
for all resources in the catalog.