(maint) Adjust rake commit check to use TRAVIS_BRANCH
On PR https://github.com/puppetlabs/puppet/pull/5280 against puppet, the travis
CI commit check is failing because a prior commit unrelated to that PR be753a4
did not adhere to the commit guidelines and the underlying check code lists
commits unreachable from master to HEAD of the PR. This commit adjust this to
check against the environment variable 'TRAVIS_BRANCH' which is populated by
travis CI with the name of the branch targeted by the PR. This should allow us
to more dynamically and correctly ascertain the correct list of commits to be
merged in a PR, and thus correctly check if any have violated our commit naming
guidelines. This commit also extends the output to include each commit that is
checked for easier debugging.

Signed-off-by: Moses Mendoza <moses@puppet.com>