(PUP-5831) Change %s to %r for Binary.new and redefine %s to UTF-8
There was a problem with the %s format as it was not symmetrical to %s
when transforming a Binary to a String. The String %s conversion is
UTF-8 based, whereas the Binary.new %s accepts the string as a raw
sequence of bytes. This is confusing and cements a bad practice that a
String is whatever you want it to be.

To fix this, a new Binary.new `%r` format is added with the meaning
"ruby raw". This format ignores the encoding and will use the verbatim
bytes. This evil construct is requires because not all strings in puppet
are UTF-8. At some point that needs to be cleaned up, but it cannot
happen until after PUP-5831 as it requires users to change functions so
that they return Binary instead of a  Ruby String with ASCII-8BIT
encoding (or lies about the content's encoding).

The `%r` format is specified as something that will be deprecated in the
future.