(PUP-6708) Cache last run report even if report_server unreachable
Prior to this commit, when the agent could not reach its report server after an
agent run, it would not save a copy of its report locally either. This is
problematic because if an agent's report server is unreachable, there is
effectively no reporting on the agent's activities during this period of
unreachability.

This behavior existed because last run report was saved via the indirector
cache - specifically, Puppet::Indirector::Indirection#save issued the terminus
save request first, then the cache save request. If a failure to send the
report to the report_server occured, the cache save was never reached.

However, unlike the last run report, the last run summary is explicitly saved
via a configurer method, save_last_run_summary, which attempts to save the last
run summary file prior to calling Puppet::Indirector::Indirection#save.

This commit moves the saving of the last run report file to a
save_last_run_report method to configurer.rb which saves the report before
sending it, following the pattern set for save_last_run_summary.

In addition to always attempting to save the report, this has an additional
benefit of explicitly managing the mode of the last_run_summary.yaml file, as
opposed to the report handler caching yaml which did not manage file mode at
all.