(PUP-6758) Fix pkgng provider condition
In cases where pkg execution fails within instances, the result returned
to prefetch is nil causing prefetch to fail poorly calling .find on a
nil object.  Here we modify the return value of instances in the event
of an exception to be an array instead of nil.  This leaves a cleaner
prefetch behavior intact, and allows the run to continue, even in the
absence of a successful package call.