(PUP-6724) Make strftime compatible with function in stdlib
This commit adds a new disptacher to the strftime function which is
backward compatible with the stdlib function with the same name.

The stdlib function also accepted a second `timezone` argument which
the new function did not provideand that was an oversight. It must be
provided to enable formatting of timezones other than UTC. The same
thing is true when parsing timestamps. A timezone must be provided to
parse a timestamp string that doesn't include a timezone unless a
default of UTC is acceptable in all cases (which it of course, isn't).
This commit therefore provides the timezone functionality needed to both
parse and format a timestamp.