(PUP-5834) Add test using apply on canned catalog
This commit contains two tests that compiles a catalog, stores it in
a file using pson, and then runs apply with the file names as input.
One tests uses `rich_data = false` and expects all resource entries
of rich data type to be converted to strings, the other test uses
`rich_data = true` and expects the rich data types to be retained.