(maint) Allow Variant of Callable in Pcore function definition
A function may have several different signatures. This commit adds the
ability to describe this using a `Variant` of `Callable` types where
otherwise just a `Callable` would be allowed.

At present, it is not possible to match each callable to a different
implementation. In order to do that, the callables must be named. Hence,
the implementation method for a function must allow all variants.