(maint) Add ability to unindent text blocks in spec-tests.
There are many times when indentation matters but yet, it's desirable
to indent the actual definition of something in an indented fashion.
Times when indentation matters is for instance when comparing output
or when defining the content of a yaml file.

This commit adds the method 'unindent' to the String class when running
spec tests.