(PUP-6807) Show pretty string when inspecting loaders
Previously, when calling inspect on a loader we would create a string
that contained the content of calling inspect on everything that is
reachable from a loader. The string formed could be larger than a
gigabyte in memory and would cause out of memory errors in production.

This replaces the default implementation of inspect with a call to #to_s
which is implemented in a number of Loader subclasses already. It also
provides a default #to_s implementation for all loaders.