(PUP-6803) report failed run on failure to generate resources
Prior to this commit, if a resource type with a defined :generate or
:eval_generate failed to generate resources (i.e., raising an exception in this
process), an exception was logged by the resource and resource generation was
skipped for that resource, but further catalog evaluation proceeded normally.

Because computing run status only took into account failed resource metrics in
determining whether to fail a run, and the resource on which :generate or
:eval_generate was called is not marked as failed, reports resulting from runs
in which resources failed to generate still claimed success, despite this
failure.

This commit adds a boolean field to
Puppet::Transaction::AdditionalResourceGenerator to track whether any resources
have raised an exception during resource generation. This value is queried by
the calling transaction instance, and is used to populate a new field of the
same name in Puppet::Transaction::Report.
Puppet::Transaction::Report#compute_status takes this value into account and
will now return `failed` if true. This commit also intentionally avoids
changing the report format by excluding the `resources_failed_to_generate` from
the yaml report.

An alternative initial approach kept a list of all resources that had failures
during resource generation and surfaced this in the report, but the decision
was made to avoid this in favor of a simple boolean to minimize resource use
and complexity, and avoid having to change the report format. Because the
resource logs an exception noting the failure, this failure information is
still surfaced in the report with a `source` populated by the resource. Thus a
user can connect their failed run status back to the exception logged in the
report without needing additional information.

In the same spirit of avoiding backwards-incompatible changes (perhaps
unfortunately) this PR does not change the resulting puppet run exit code if
resources fail to generate - all things being equal this scenario will still
return 0.

Signed-off-by: Moses Mendoza <moses@puppet.com>