(PUP-6474) Search for locale files in system
Previously, we were only looking in the current repo for locale
files. This was fine for development, but is not where the files will be
installed in a package. This adds the path where the files will end up
in packaging, and does a simple check to see if it looks like we're in a
developer environment.