(PUP-1476) Include puppet and ruby versions in User-Agent
Previously, puppet did not specify an HTTP User-Agent string, so ruby
would send the default "Ruby". This made it difficult to debug and
interpret server-side logs when different versions of puppet agents
connect.

This commit adds a puppet setting `http_user_agent` which includes the
puppet and ruby versions. The default value is of the form:

    Puppet/4.7.0 Ruby/2.1.8-p440 (x86_64-darwin15.0)

The string is exposed as a setting for cases where puppet is being used
as a library, and the enclosing application may need to override that.

Puppet's HTTP connection code will send the User-Agent header with the
above value for all requests. The value is taken from the Puppet Module
Tool code. However, the strings are not exactly the same, because the
PMT prefixes the string with additional information:

    PMT/1.1.1 (v3; Net::HTTP) Puppet/4.7.0 Ruby/2.1.8-p440 (x86_64-darwin15.0)