(PUP-6827) fix test to meet password complexity reqs
Prior to this commit, the test '#6857: redact password hashes when applying in
noop mode' would fail on a host that had 'Password must meet complexity
requirements' set to true, because puppet was applying user management
manifests containing passwords that did not meet said requirements. This commit
updates the test to meet the complexity requirements (documented here:
https://technet.microsoft.com/en-us/library/hh994562(v=ws.11).aspx), allowing tests to pass.

Assumedly prior hosts in the acceptance testing systems were either
pre-configured with this requirement disabled or it was disabled at test time
for other tests (the latter of which seems unlikely).

Signed-off-by: Moses Mendoza <moses@puppet.com>