(PUP-6789) Allow capabilties to be consumed across environments.
Previously the capability finder would only search inside the
environment of the current catalog for capability resources to consume.
Now the capability finder will first search for a matching capability
across all environments and only limit it's search to a single
environment if a conflict is found. This allows users to test new code
on part of an application by moving nodes into a different environment
as long as they use globally unique capability titles.