(PUP-6780) Avoid endless recursion when debugging the Lookup Explainer
The lookup explainer relied on the method to_s to provide its explanation.
This was not ideal since that method is called by the debugger. This commit
therefore adds a specific #explain method to produce the explanation and
changes the places where to_s was called.

Another minor improvement was also made. The handling of abitrary texts
was moved to the most general class in the explainer tree structure so that
the root node may contain texts.