(maint) Add spec coverage for `puppet config set`
Prior to this commit, there was nothing in specs (at least not that I could
find) that exercised and tested the :set action to `puppet config`. Prior
changes appear to have added testing only in the Puppet::Settings namespace.
This commit adds new specs that validate that `puppet config set`:
  1) actually opens the correct config file
  2) creates the config file if it doesn't exist
  3) sets the correct value in the config file
  4) sets the value in the correct config section if supplied

Signed-off-by: Moses Mendoza <moses@puppet.com>