(PUP-6188) use Puppet::FileSystem in ssl
This commit updates the various calls to File.read throughout Puppet::SSL to
instead call their Puppet::FileSystem counterpart, to further centralize
control of file system access in our library with the advantages it affords.
The second aspect of this PR is to ensure we are specifying :encoding =>
'UTF-8' wherever we do access files, so that we can appropriately handle
content outside of the ASCII character range (for example, certs with subject
fields containing special characters, ie locations).

Signed-off-by: Moses Mendoza <moses@puppet.com>