(PUP-7134) Use Puppet::FileSystem in ssl
This commit updates the various calls to File.read throughout
Puppet::SSL to instead call their Puppet::FileSystem counterpart, to
further centralize control of file system access in our library with
the advantages it affords.

The second aspect of this commit specifes the :encoding attribute
wherever files are accessed, so that a character set is explicitly
enforced that is compatible with SSL related files on disk, such as
.pem, .csr, etc - at relevant call sites it's documented what IETF or
other specs cover the encoding of a given document.

Signed-off-by: Moses Mendoza <moses@puppet.com>
               Ethan J. Brown <ethan@puppet.com>