(PUP-6847) Omit whitespace from provider keyword in network_device
Prior to this update, any trailing whitespace left in a device
configuration file will cause `puppet device` to fail with
a cryptic message because it is using the string to construct
the path for an require statement.  Since require statements
should never have spaces in them, it makes more sense to have the
regex only accept the non-whitespace characters it finds in the
confiugration file.