(PUP-6188) use UTF-8 for files in util/filetype.rb
This commit updates invocations of Tempfile and IO to open with 'UTF-8'
encoding, as well as replaces a call to File.read with Puppet::FileSystem.read
with the UTF-8 encoding as well. This is part of a larger effort to ensure
Puppet properly handles characters outside the ASCII range as we move towards
localization.

Signed-off-by: Moses Mendoza <moses@puppet.com>