(PUP-6871) Fix transaction persistent unit test for non English Windows
Previously the unit test for transaction/persistence checked whether an error
was raised with specific text.  However the text was in English and the test
would fail if used on localized versions of Windows, such as French, where the
message is French.  This commit removes the text in the expectation so that it
will behave correctly on all language versions of Windows and instead uses the
error code of 5 to detect an Access Denied error.