(PUP-6897) Fix for agent mishandling in pcore-gen and cor-change tests
Previously, the pcore_generate_env_isolation.rb and
corrective_change_via_puppet.rb tests were failing when run in a CI
configuration with more than one agent in it.  In each case, the test
changed some state after an agent was run on the first node which was
unintentionally picked up as the agent was run on the second node.  This
commit changes both tests to perform all related actions on each of the
nodes before any state changes are made.