(PUP-6898) Fix sensitive_data_type test validation of syslog output
Previously, the sensitive_data_type test, when attempting to validate
the presence / absence of expected content in the syslog, performed
assertions against the full Beaker logger instead.  This would cause
assertions looking for the absence of sensitive content to fail for the
second agent run performed because the assertions themselves would
appear in the logger output.

This commit changes the syslog assertion logic to inspect just the
content of the syslog - without bringing in the rest of the current
Beaker logger.  The commit also only runs through the assertions which
confirm that sensitive data does not appear in the syslog and not the
assertions which confirm that other debug messages from the agent run
appear in the syslog.  This is done because the agent is run with a
"--debug" flag, which should cause agent log messages to go to the
console and not to the syslog anyway.