(PUP-6856) Catch :no_such_key during interpolation sub_lookup
An interpolation that uses dotted keys that in turn results in a
`:no_such_key` being thrown (because the key is missing), would not catch the
throw and hence, yield strange results. This commit ensures that the thrown
symbol `:no_such_key` is caught so that the call instead results in an empty
string.