(PUP-6834) Ensure that input to MergeStrategy deep does not mutate
Before this commit, the primary source for a deep merge was dup'ed to avoid
that the deep merge would mutate it (it mutates the primary source). A dup
is however not enough when the source is a nested hash. The dup will only
perform a shallow copy. This commit ensures that the source is deep copied
instead.