(PUP-6911) Do uptime comparisons as integers in var-refresh test
Previously, the `variables_refreshed_each_compilation` acceptance test
could fail if the "earlier" uptime value that the test was comparing for
an agent run was lexicographically greater than the "later" uptime value.
For example, an earlier uptime of "9999" would be considered greater
than a later uptime of "10002" and, therefore, fail.

This commit changes the test to convert the uptimes to Integers before
comparing them so that the comparisons are done numerically rather than
as strings.