(PUP-7104) Differentiate between base64 and binary data
The current implementation would always serialize a Binary as a base64
encoded String. This is inneficient when using MsgPack.

This commit introduces a `supports_binary?` method, renames the existing
extension from `BINARY` to `BASE64` and then adds an additional extension
named `BINARY`. The choice between the two is controlled by the new method.