(PUP-7104) Fix for serialization of Sensitive with rich data.
The current implementation of Sensitive serialization was broken. It assumed
that the same packer was used for both normal data and extension data and
made attempts to write rich data directly into the extension buffer. While
fine for JSON, that approach doesn't work at all for MsgPack.

This commit changes the serialzation so that a SensitiveStart marker is
written using an extension point and then the actual wrapped object is
written as a normal object.