(PUP-6915) Specify log destination for sensitive data type test
By default, puppet logs to different syslog directories depending on the
platform it is running on. When Fedora recently changed how to access
system logs, this test broke, since it was explicitly looking for
specific default log locations on various platforms. This commit refactors the test
to specify a temp file to receive the logs, thereby avoiding brittle
platform-specific code.