(PUP-6915) Fix sensitive data type test on Solaris 10
A recent refactor of the sensitive data type test to use temp files for
logging instead of default system directories allowed us to run this
test on more platforms. However, the `tail -n` command we were using is
not supported on Solaris 10. This updates the command to drop the `-n`,
which still accomplishes the same result, and works on all platforms.